+++ ua.rd.project4/src/main/java/ua/rd/project4/model/holders/CarHolder.java
@@ -2,8 +2,10 @@ package ua.rd.project4.model.holders;
 
 import ua.rd.project4.domain.Car;
 
+import java.math.BigDecimal;
+
 public class CarHolder extends Car   {
-    public CarHolder(String model, String color, CarType carType, String registrationNumber, String description, int price, int rentPricePerDay) {
+    public CarHolder(String model, String color, CarType carType, String registrationNumber, String description, BigDecimal price, BigDecimal rentPricePerDay) {
         super(model, color, carType, registrationNumber, description, price, rentPricePerDay);
     }
 }
